[od50] Fixed compile problems.
Summary: http://intranet:8080/cdash/viewBuildError.php?type=1&buildid=47403

Test Plan: Visually inspected only.

Reviewers: aneesh

Reviewed By: aneesh

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D5492

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@36286 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/AttributeEngine/attribdatapack.cc
@@ -849,7 +849,8 @@ bool FlatDataPackZAxisTransformer::doWork(
 	    continue;
 
 	SampledFunctionImpl<float,ValueSeries<float> > inputfunc(
-		arr1dslice, inpzrg.nrSteps()+1, inpzrg.start, inpzrg.step );
+		arr1dslice, inpzrg.nrSteps()+1, (float) inpzrg.start,
+		(float) inpzrg.step );
 	inputfunc.setHasUdfs( true );
 	inputfunc.setInterpolate( interpolate_ );
